[Wave] add excalidraw visualization of wave compilation pipeline (#777)
This adds a visualization of the pipeline Wave uses for compilation.

The .png file was created with [excalidraw](https://excalidraw.com/) and
contains all info to further modify it.
e.g. I use the [VSCode
extension](https://marketplace.visualstudio.com/items?itemName=pomdtr.excalidraw-editor)
to modify `*.excalidraw.png` files. This allows having the image and
source to render/modify it in a single file rather than maintaining both
`pipeline.excalidraw` and `pipeline.png` (or having no source for it
entirely).

Signed-off-by: Martin Lücke <martin@brium.ai>